server/requirements.txt
Django==4.2.10
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
django-filter==23.5
Pillow==10.1.0
django-environ==0.11.2
django-two-factor-auth==1.15.5
django-otp==1.2.2
qrcode==7.4.2
python-dateutil==2.8.2
dj-rest-auth==4.0.1
django-allauth==0.57.0
gunicorn==21.2.0
whitenoise==6.6.0
drf-yasg==1.21.7
